feat(loadout): add trinket equipment system and unit config lookups
Introduce trinket ownership and equipping alongside existing weapons,
with new endpoints to fetch a single unit config and list all owned
weapons/trinkets. Also switch damage stat modifiers from float
multipliers to int percentages for precision.

+++ Models/EquipmentDefinition.cs
@@ -35,6 +35,6 @@ public class EquipmentStatModifiers
     public int MaxHP { get; set; }
     public int Speed { get; set; }
     public int MaxSkillPoint { get; set; }
-    public float DamageMultiplier { get; set; }
-    public float DamageReduction { get; set; }
+    public int DamageMultiplier { get; set; }
+    public int DamageReduction { get; set; }
 }
